Update synology_snmp.sh
updated the time management of the script. previously it would increment a counter every script execution and save the counter to the disk. this resulted in disk writes every 60 seconds. 

now the script saves the linux time stamp when an email notification is sent and only saves the data to disk after the email is successfully sent. This will prevent the script from constantly writing data to disk every 60 seconds.